Separate simulation sound interpretation from frame orchestration
